refactor(view): answer 패키지 추가
