+++ src/main/java/janggi/game/JanggiGame.java
@@ -6,8 +6,8 @@ import janggi.setting.CampType;
 import janggi.setting.GameState;
 import janggi.setting.PieceAssignType;
 import janggi.view.GameInputOutput;
-import janggi.view.GameMenuAnswer;
-import janggi.view.TurnMenuAnswer;
+import janggi.view.answer.GameMenuAnswer;
+import janggi.view.answer.TurnMenuAnswer;
 import java.util.List;
 import java.util.Optional;
