+++ src/main/java/janggi/view/GameInputOutput.java
@@ -6,6 +6,8 @@ import janggi.game.MovePieceCommand;
 import janggi.setting.CampType;
 import janggi.setting.PieceAssignType;
 import janggi.value.Position;
+import janggi.view.answer.GameMenuAnswer;
+import janggi.view.answer.TurnMenuAnswer;
 import java.util.List;
 import java.util.Optional;
