+++ src/main/java/janggi/view/answer/GameMenuAnswer.java
@@ -1,20 +1,20 @@
-package janggi.view;
+package janggi.view.answer;
 
 import java.util.Arrays;
 
-public enum TurnMenuAnswer {
+public enum GameMenuAnswer {
     ONE("1"),
     TWO("2"),
-    THREE("3");
+    QUIT("q");
 
     private final String command;
 
-    TurnMenuAnswer(String command) {
+    GameMenuAnswer(String command) {
         this.command = command;
     }
 
-    public static TurnMenuAnswer parse(final String number) {
-        return Arrays.stream(TurnMenuAnswer.values())
+    public static GameMenuAnswer parse(final String number) {
+        return Arrays.stream(GameMenuAnswer.values())
                 .filter(answer -> answer.getCommand().equals(number))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("[ERROR] 적절하지 않은 입력값입니다."));
