+++ src/main/java/janggi/view/answer/TurnMenuAnswer.java
@@ -1,20 +1,20 @@
-package janggi.view;
+package janggi.view.answer;
 
 import java.util.Arrays;
 
-public enum GameMenuAnswer {
+public enum TurnMenuAnswer {
     ONE("1"),
     TWO("2"),
-    QUIT("q");
+    THREE("3");
 
     private final String command;
 
-    GameMenuAnswer(String command) {
+    TurnMenuAnswer(String command) {
         this.command = command;
     }
 
-    public static GameMenuAnswer parse(final String number) {
-        return Arrays.stream(GameMenuAnswer.values())
+    public static TurnMenuAnswer parse(final String number) {
+        return Arrays.stream(TurnMenuAnswer.values())
                 .filter(answer -> answer.getCommand().equals(number))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("[ERROR] 적절하지 않은 입력값입니다."));
